ConnectionStatusクラス、checkConnectivityメソッドの定義場所を変更
